requirements.txt
Flask
boto3
cognitojwt
python-dotenv
